improve test structure
