+++ tests/smoke/redis_test.py
